alarms/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
uvicorn==0.34.3
requests==2.32.3
